pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sky-governance-zep"
version = "0.1.0"
description = "Temporal knowledge graph over MakerDAO/Sky governance history, built on ZEP Cloud"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "zep-cloud>=2.0.0,<3.0.0",
    "requests>=2.31.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-mock>=3.12.0,<4.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["governance"]
